Add manage_limits_d_dir bool class param
Without this patch the use of the limits module in conjunction with the
[pam][pam] module causes the following error:

    Failed to apply catalog: Cannot alias File[limits_d] to ["/etc/security/limits.d"] at
    /etc/puppet/ghoneycutt-modules/pam/manifests/limits.pp:62;
    resource ["File", "/etc/security/limits.d"] already declared at
    /etc/puppet/environments/development/modules/limits/manifests/init.pp:15

This patch addresses the issue by allowing the end user to disable management of
the limits.d directory.  This is desirable because the end user may still wish
to manage limits using the `limits::entries` Hiera data value in conjunction
with the pam module.

[pam]: https://git.io/v6qAH